Data Processing / Data Sets for FE & MODELS
These are the data sets we will be using to create our input and output features as well as run the ML or SL models on, and maybe neural nets. Please let me know if there are any questions. merged_data.csv has all of the feature/factor data from VIX, CAPE, FRED, and Macro_variables datasets from the Box. intialData.py is the python file that made this happen & will  be updated soon depending on how much preprocessing needs to be done.